+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 5.2.0-beta.1 (Unreleased)
 ### Feature Added
 - We are now targeting the service's v3.2-preview.1 API as the default instead of v3.1.
-- Added support for the `Extractive Summmarization` new feature.
+- Added support for Extractive Summarization actions through the `ExtractSummaryAction` type.
 
 ## 5.1.0 (2021-07-08)
 ### Feature Added